shared/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "service-core"
version = "1.1.0"
description = "Core infrastructure for Prelude Platform services (pool, auth, tenant routing, email)"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "anthropic>=0.80.0",
    "httpx>=0.27.0",
    "openai>=1.6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
]

[tool.setuptools.packages.find]
include = ["service_core*", "email_core*"]
